+++ tests/features/src/src5.v
@@ -14,7 +14,7 @@
     assign data_o = tmp;
 
     module4 module4_inst (
-        (* some_attr = "true" *) .clk_i (clk_i),
+        (* some_attr = "true" *) .clk_i (clk_local),
         .data_o (from_module4)
     );